Thunks: Disable 32-bit thunks unless a config option is set
This will be necessary before #3487 is merged.
Once #3487 is merged then 32-bit Vulkan thunk libraries is built which
results in crashing in every 32-bit vulkan/dxvk game.

Explicitly disable the thunking until this config option is set. Once
X11 is as stable on 32-bit as 64-bit then this option can get removed.

For testing purposes it's easy to set the environment variable
`FEX_THUNKS32BITENABLED=1`