CPUInfo: Optimize CalculateNumberOfCPUs
Caches the result for future requests since we call this multiple times.

Additionally instead of using fmt::format_to_n, it instead uses
std::to_chars for just the index at the end of the filename.

This causes the function to go from around 146k ns to 55k ns, nearly
1/3rd the amount of time to execute.