FileManagement: Trace symlink paths across rootfs
Non-rootfs paths might contain a path component that links to the
rootfs. This is the case with Wine's `dosdevices/z:` which links to /.
To handle this, we need to trace the path component by component and
resolve each potential symlink (except the last one, which is already
handled by the existing logic if FollowSymlink is true).

Signed-off-by: Asahi Lina <lina@asahilina.net>